refactor: seasonrecapcard is now more general use of component called seasoncard
