+++ posts/season-recaps/index.md
@@ -7,5 +7,5 @@ next: false
 # Kausiyhteenvedot
 
 <div class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-3 gap-4 p-4 mt-8">
-  <SeasonRecapCard title="Kanaliiga CS2 Season 1" image="../images/kanaliiga.png" description="Kausi päättyi ja on aika kerrata, että miten Triplanilaiset pärjäsivät tällä kaudella..." read-more-link="./kanaliiga-cs2-season-1"/>
+  <SeasonCard title="Kanaliiga CS2 Season 1" image="../images/kanaliiga.png" description="Kausi päättyi ja on aika kerrata, että miten Triplanilaiset pärjäsivät tällä kaudella..." read-more-link="./kanaliiga-cs2-season-1"/>
 </div>
